+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### 2.6.9 - 2024-10-13
+* Estonian payment method list layout updated
+
 ### 2.6.8 - 2024-08-28
 * 274258: Add instructions link to settings
 * 274259: Display error message when payment methods are not available